requirements.txt
selenium==4.15.2
python-dotenv==0.20.0
pymongo==4.3.3
apache-airflow==2.5.1
google-api-python-client==2.97.0